Routing optimizations, Zip-Tie holes
Numerous routing changes / optimizations were made to help make the
best use of the ground and power planes for both power/gnd delivery
as well as isolation. On that note, the isolation on both planes
was lowered to 12 mil.

Added two more zip-tie holes. The cuts were not showing up on
the OshPark previews so I switched to using a line, which does show
up. Not enitrely show how well they exactly they will come out,
but I think the clearance is enough.